backend/requirements.txt
setuptools>=69.0.0
wheel>=0.42.0
pip>=24.0.0

fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.9.0
websockets>=12.0

SpeechRecognition==3.10.0
pydub>=0.25.1
edge-tts>=6.1.9
gTTS>=2.4.0

deep-translator>=1.11.4
google-generativeai>=0.3.0

numpy>=1.26.0
